+++ nanonis_qcodes_controller/qcodes_driver/__init__.py
@@ -1,6 +1,8 @@
 from .extensions import (
+    ActionSpec,
     ParameterSpec,
     ScalarParameterSpec,
+    load_action_specs,
     load_parameter_specs,
     load_scalar_parameter_specs,
 )
@@ -8,8 +10,10 @@ from .instrument import QcodesNanonisSTM
 
 __all__ = [
     "QcodesNanonisSTM",
+    "ActionSpec",
     "ParameterSpec",
     "ScalarParameterSpec",
+    "load_action_specs",
     "load_parameter_specs",
     "load_scalar_parameter_specs",
 ]